challenge-codes-main/challenge-codes-main/prototipo/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
